NTR-19 JNDI Profile
Add spring profile for JNDI datasource.

Currently defaults to non-JNDI to allow time for
folks to configure their tcServer.

To create a JNDI datasource, add this to your
context.xml...

    <Resource
        name="jdbc/nanodb"
        type="javax.sql.DataSource"
        auth="Container"
        driverClassName="org.postgresql.Driver"
        maxActive="20"
        maxIdle="10"
        maxWait="-1"
        username="nanotrader"
        password="nanotrader"
        url="jdbc:postgresql://nanodbserver:5432"
        testWhileIdle="true"
        testOnBorrow="true"
        testOnReturn="false"
        validationQuery="SELECT 1"
        validationInterval="30000"
        timeBetweenEvictionRunsMillis="5000"
        initialSize="10"
        removeAbandonedTimeout="60"
        removeAbandoned="true"
        logAbandoned="true"
        minEvictableIdleTimeMillis="30000"
        jmxEnabled="true"
       />

(or the equivalent for SqlFire).

When ready, change the web.xml to use 'jndi' instead
of 'default' in the 'spring.active.profiles' init
param.

This needs to be done in the ...services and
...async-services projects.